refactor(bin): moved code out of bin.js
